Split storage.js into one file per class.
